implement song voting sys

+++ src/main/java/io/github/therosgit/resonate/ochestrator/repository/FingerprintRepository.java
@@ -7,6 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import io.github.therosgit.resonate.ochestrator.domain.Fingerprint;
 
 public interface FingerprintRepository extends JpaRepository<Fingerprint, Long> {
-
-    List<Fingerprint> findByHashIn(List<Integer> hashes);
+    List<Fingerprint> findByHashIn(List<Long> hashes);
 }
